bak/src/requirements.txt
redis
paho-mqtt
pandas
taospy
pyodbc
Flask
mysql-connector-python